apps/api/requirements.txt
# Core dependencies
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
pandas>=2.2.0
openpyxl>=3.1.2
boto3>=1.34.0
python-dotenv>=1.0.0
supabase>=2.3.0
openai>=1.10.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# ML dependencies (optional, install with: pip install -r requirements.txt && pip install torch torchvision torchaudio)
# torch>=2.2.0
# torchvision>=0.17.0
# torchaudio>=2.2.0

# Development dependencies (optional)
# pytest>=7.4.0
# pytest-asyncio>=0.23.0
# httpx>=0.26.0
